+++ src/runtime/transformation-dom-particle.ts
@@ -1,0 +1,44 @@
+/**
+ * @license
+ * Copyright (c) 2017 Google Inc. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * Code distributed by Google as part of this project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+'use strict';
+import { DomParticle } from './dom-particle.js';
+// Regex to separate style and template.
+const re = /<style>((?:.|[\r\n])*)<\/style>((?:.|[\r\n])*)/;
+/** @class TransformationDomParticle
+ * Particle that does transformation stuff with DOM.
+ */
+export class TransformationDomParticle extends DomParticle {
+    getTemplate(slotName) {
+        // TODO: add support for multiple slots.
+        return this._state.template;
+    }
+    getTemplateName(slotName) {
+        // TODO: add support for multiple slots.
+        return this._state.templateName;
+    }
+    render(props, state) {
+        return state.renderModel;
+    }
+    shouldRender(props, state) {
+        return Boolean((state.template || state.templateName) && state.renderModel);
+    }
+    renderHostedSlot(slotName, hostedSlotId, content) {
+        this.combineHostedTemplate(slotName, hostedSlotId, content);
+        this.combineHostedModel(slotName, hostedSlotId, content);
+    }
+    // abstract
+    combineHostedTemplate(slotName, hostedSlotId, content) { }
+    combineHostedModel(slotName, hostedSlotId, content) { }
+    // Helper methods that may be reused in transformation particles to combine hosted content.
+    static propsToItems(propsValues) {
+        return propsValues ? propsValues.map(({ rawData, id }) => Object.assign({}, rawData, { subId: id })) : [];
+    }
+}
+//# sourceMappingURL=transformation-dom-particle.js.map
